Background
Background
Final project in Advanced software development course as part of our B.Sc. Computer Science studies. The project presents a visual analysis of flight control details.

The Flight simulator can connect to the FlightGear software according to TCP protocol, Our simulator knows to tell us where anomalies in the values ​​of the interrogated interrogation were detected given a normal flight file and another interrogated flight file. We have developed for the investigation of the anomalies anomalies algorithms.

The app does not know the anomaly algorithms in advance but loads them as a plug-in, And so the user can build his own anomaly algorithms and add them himself.